// UI/ClientApp/src/app/Core/Models/practicalModel.ts
export class PracticalModel
 {
  practicalId: number;
  practicalName: string;
  category: number;
  defination: string;
  comment: string;
  technology: string;
  experienceLb: number;
  experienceUb: number;
  versionNumber: number;
  fileName: string;
  fileType: number;
  categoryName: string;
  fileTypeName: string;
  pracDocumentName: string;
  refDocumentName: string;
}